refactor: consolidate Markdown semantic partials [ci]
